moved setup provider to cluster
Makes more sense there
