+++ internal/cli/cli.go
@@ -35,7 +35,7 @@ func BuildCli() {
 	providerFlag := &cli.StringFlag{Name: "provider", Usage: "Select k8s provider (GCP only for now) "}
 
 	keyPath := &cli.StringFlag{Name: "keyPath", Usage: "SA key path"}
-	setupEnvironment := setupEnvironmentCmd(providerFlag, projectFlag, keyPath)
+	setupEnvironment := setupEnvironmentCmd(providerFlag, projectFlag, environmentFlag, keyPath, dryrRunFlag)
 	dockerSetup := dockerSetupCmd(projectFlag, environmentFlag)
 	dockerBuild := dockerBuildCmd(projectFlag)
 
@@ -285,7 +285,7 @@ func createClusterCmd(projectFlag *cli.StringFlag, environmentFlag *cli.StringFl
 
 			project := c.String("project")
 			env := c.String("env")
-			dryrunflag := c.Bool("dry-run")
+			dryrun := c.Bool("dry-run")
 			provider := c.String("provider")
 			key := c.String("keyPath")
 
@@ -296,18 +296,9 @@ func createClusterCmd(projectFlag *cli.StringFlag, environmentFlag *cli.StringFl
 				gcp.SetKeyEnvVar(key)
 			}
 			//TODO: pass-in variables
-			err := cluster.CreateCluster(project, env, zone, dryrunflag)
+			err := cluster.CreateCluster(provider, project, env, zone, dryrun, key)
 			if err != nil {
 				log.Fatalf("Could not create cluster in project: %v. Error: %v", project, err)
-			}
-			if !dryrunflag {
-				// Set local env kube for local connectivity to new cluster
-				log.Printf("Setting kubectl to work with new cluster: %v", project+"-"+env)
-				SetupProvider(provider, zone, project+"-"+env, project, key, dryrunflag)
-				cluster.InstallHelm()
-			} else {
-				log.Printf("Authenticating only due to --dry-run: %v", project+"-"+env)
-				SetupProvider(provider, zone, project+"-"+env, project, key, dryrunflag)
 			}
 
 			return nil
