+++ internal/cluster/provider.go
@@ -1,4 +1,4 @@
-package cli
+package cluster
 
 import (
 	"fmt"
